+++ pwsh/dev/functions/getSubscriptions.ps1
@@ -6,11 +6,31 @@ function getSubscriptions {
     $method = 'GET'
     $requestAllSubscriptionsAPI = AzAPICall -AzAPICallConfiguration $azAPICallConf -uri $uri -method $method -currentTask $currentTask
 
+    $script:htAllSubscriptionsFromAPI = @{}
+    $script:htSubscriptionsFromOtherTenants = @{}
+
     Write-Host " $($requestAllSubscriptionsAPI.Count) Subscriptions returned"
     foreach ($subscription in $requestAllSubscriptionsAPI) {
-        $script:htAllSubscriptionsFromAPI.($subscription.subscriptionId) = @{}
-        $script:htAllSubscriptionsFromAPI.($subscription.subscriptionId).subDetails = $subscription
+
+        # #test
+        # if ($subscription.subscriptionId -eq 'GUID') {
+        #     Write-Host "  Finding: $($subscription.displayName) ($($subscription.subscriptionId)) belongs to foreign tenant '$($subscription.tenantId)' - AzGovViz: excluding this Subscripion" -ForegroundColor DarkRed
+        #     $script:htSubscriptionsFromOtherTenants.($subscription.subscriptionId) = @{}
+        #     $script:htSubscriptionsFromOtherTenants.($subscription.subscriptionId).subDetails = $subscription
+        #     continue
+        # }
+
+        if ($subscription.tenantId -ne $azAPICallConf['checkcontext'].tenant.id) {
+            Write-Host "  Finding: $($subscription.displayName) ($($subscription.subscriptionId)) belongs to foreign tenant '$($subscription.tenantId)' - AzGovViz: excluding this Subscripion" -ForegroundColor DarkRed
+            $script:htSubscriptionsFromOtherTenants.($subscription.subscriptionId) = @{}
+            $script:htSubscriptionsFromOtherTenants.($subscription.subscriptionId).subDetails = $subscription
+        }
+        else {
+            $script:htAllSubscriptionsFromAPI.($subscription.subscriptionId) = @{}
+            $script:htAllSubscriptionsFromAPI.($subscription.subscriptionId).subDetails = $subscription
+        }
     }
+    Write-Host " $($htAllSubscriptionsFromAPI.Keys.Count) Subscriptions relevant"
 
     $endGetSubscriptions = Get-Date
     Write-Host "Getting all Subscriptions duration: $((New-TimeSpan -Start $startGetSubscriptions -End $endGetSubscriptions).TotalSeconds) seconds"